+++ versions.tf
@@ -3,7 +3,7 @@ terraform {
     aws = {}
     clumio = {
       source  = "clumio-code/clumio"
-      version = ">=0.16.1, <0.18.0"
+      version = ">=0.17.0, <0.19.0"
     }
   }
 }